broker-manager/requirements.txt
fastapi==0.92.0
requests==2.28.2
uvicorn==0.20.0
psycopg2-binary==2.9.5
